+++ jmock-junit5/src/test/java/org/jmock/junit5/testdata/jmock/acceptance/BaseClassWithMockery.java
@@ -1,4 +1,4 @@
-package org.jmock.junit5.testdata;
+package org.jmock.junit5.testdata.jmock.acceptance;
 
 import org.jmock.junit5.JUnit5Mockery;
 import org.junit.jupiter.api.extension.RegisterExtension;
