+++ jmock-junit5/src/test/java/org/jmock/junit5/testdata/jmock/acceptance/DerivedJUnit5TestThatDoesNotSatisfyExpectations.java
@@ -1,4 +1,4 @@
-package org.jmock.junit5.testdata;
+package org.jmock.junit5.testdata.jmock.acceptance;
 
 import org.jmock.Expectations;
 import org.jmock.auto.Mock;
